+++ example/android/app/src/main/kotlin/com/example/islamic_text_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.example.islamic_text_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()